utilitiesService/requirements_windows.txt
aniso8601==8.1.0
attrs==20.3.0
cffi==1.14.4
click==7.1.2
cycler==0.10.0
Flask==1.0.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
flask-restplus==0.13.0
flask-swagger==0.2.13
gevent==20.5.0
gevent==20.5.0
greenlet==0.4.17
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
netifaces==0.10.9
numpy==1.19.1
pandas
Pillow==7.2.0
py-eureka-client==0.6.0
pycparser==2.20
pymongo==3.11.2
pyparsing==2.4.7
pyprg==0.1.1b7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2020.4
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
torch==1.4.0+cpu # On Windows CPU only ->  pip install torch==1.4.0+cpu torchvision==0.5.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.5.0+cpu
urllib3==1.24.1
Werkzeug==0.14.1
zipp==3.4.0
scikit-image==0.17.2
connexion==2.7.0
IPython==7.16.1
flask_cors==3.0.10